Rename tmp module to file_mgmt in prep for added func.
